Wrap all message converters from spring-data-rest with a multipart-supporting message converter
In spring-data-rest 4.4.0, the logic has changed and now
`RepresentationModel`is being passed to `canRead()` instead of
`PersistentEntityResource`.

To support both versions of spring-data-rest for a while, we can wrap all
message converters with one that supports multipart and converts it to
JSON. This also allows users to configure their own specific HttpMessageConverters
and have it automatically work with multipart as well.