Simplify encryption for byte-range requests
Because an actual compliant InputStream is now returned from
the backing store, encryption can be handled simpler by just running the
cipher for the whole range.

Since encryption is not authenticated, and CTR mode is byte-addressable,
this just works. The zero-bytes before/after the actual data are
decrypted as garbage, but they will be cut out when streaming the
response back to the user.