Fix S3 PartialContentInputStream to be compliant to InputStream specification (#2217)
* Make S3 PartialContentInputStream spec-compliant

The PartialContentInputStream was a minimal implementation that worked
well when used directly with the spring byte-range support.

However, using the InputStream variant in any other place (like wrapping it for decryption),
the minimal implementation breaks down, requiring workarounds when
handling the encryption.
Using an implementation that is fully compliant with the InputStream
specifications makes implementation of a wrapping InputStream easier,
as no special provisions need to be made for a read method that
immediately reads the byte-range (instead of empty space before the
start)

* Simplify encryption for byte-range requests

Because an actual compliant InputStream is now returned from
the backing store, encryption can be handled simpler by just running the
cipher for the whole range.

Since encryption is not authenticated, and CTR mode is byte-addressable,
this just works. The zero-bytes before/after the actual data are
decrypted as garbage, but they will be cut out when streaming the
response back to the user.

* Make range-request not aligned to an AES block boundary

Although the testcase works when aligned on a block boundary, the test is more comprehensive when it is not aligned

* Add back block skipping for offsets

Skipping decrypting blocks that do not have data and will be skipped anyways is wasted work.
Reintroduce the technique to calculate IV for a certain block offset to save decryption work for larger offsets

* Use long for file offsets

Using an int limits the offset to about 2GB, which might not be sufficient for very large files.
Using a long results in an offset of 9EB, which should be sufficient.